Split out Motors and Motor for clarity. Motors is the main access point of the API, Motor is the base class.
