+++ Grbl_Esp32/src/grbl.h
@@ -57,7 +57,7 @@
 #include "serial.h"
 #include "Pins.h"
 #include "Spindles/SpindleClass.h"
-#include "Motors/Motor.h"
+#include "Motors/Motors.h"
 #include "stepper.h"
 #include "jog.h"
 #include "inputbuffer.h"
